Revert "Move FollowButton to blocks"
