+++ client/devdocs/design/blocks.jsx
@@ -20,7 +20,7 @@ import SearchCard from 'components/search-card';
 import CreditCardForm from 'blocks/credit-card-form/docs/example';
 import AuthorSelector from 'blocks/author-selector/docs/example';
 import CommentButtons from 'blocks/comment-button/docs/example';
-import FollowButton from 'blocks/follow-button/docs/example';
+import FollowButton from 'components/follow-button/docs/example';
 import LikeButtons from 'blocks/like-button/docs/example';
 import PostSchedule from 'components/post-schedule/docs/example';
 import PostSelector from 'my-sites/post-selector/docs/example';
